+++ plugins/supervisor/__init__.py
@@ -1,5 +1,0 @@
-"""Supervision plugin: restart-on-failure policy over kernel failure signals."""
-
-from plugins.supervisor.supervisor_plugin import SupervisorPlugin
-
-__all__ = ["SupervisorPlugin"]